feat(2fa): set up user session management
ref: MANAGER-13716
