+++ packages/manager/apps/account-disable-2fa/src/App.tsx
@@ -6,8 +6,6 @@ import { RouterProvider, createHashRouter } from 'react-router-dom';
 import '@ovhcloud/ods-theme-blue-jeans';
 import queryClient from './query.client';
 import { Routes } from './routes/routes';
-import { User } from '@/components/User/context';
-import UserProvider from '@/components/User/provider';
 
 odsSetup();
 
@@ -15,17 +13,11 @@ const router = createHashRouter(Routes);
 
 const Router = () => <RouterProvider router={router} />;
 
-type AppProps = {
-  user: User;
-};
-
-export default function App({ user }: AppProps) {
+export default function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <UserProvider user={user}>
-        <Router />
-        <ReactQueryDevtools initialIsOpen={false} />
-      </UserProvider>
+      <Router />
+      <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   );
 }
