--- packages/manager/apps/account-disable-2fa/src/components/User/useUser.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-import userContext, { UserContext } from '@/components/User/context';
-
-const useUser = (): UserContext => {
-  return useContext(userContext);
-};
-
-export default useUser;
