+++ packages/manager/apps/account-disable-2fa/src/context/User/useUser.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import userContext, { UserContext } from '@/context/User/context';
+
+export default (): UserContext => {
+  return useContext(userContext);
+};
